Another SafeDivide cast warning fix for MSVC.
Eliminates MSVC warnings related to SafeDivide by using the same type for both return and input and requires an explicit cast where needed.

+++ src/gpgmm/utils/WindowsTime.cpp
@@ -29,7 +29,7 @@ namespace gpgmm {
             LARGE_INTEGER curTime;
             const bool success = QueryPerformanceCounter(&curTime);
             ASSERT(success);
-            return SafeDivide(curTime.QuadPart, GetFrequency());
+            return static_cast<double>(SafeDivide(curTime.QuadPart, GetFrequency()));
         }
 
         void StartElapsedTime() override {
